requirements.txt
appnope==0.1.0
backcall==0.1.0
boto3==1.7.22
botocore==1.10.22
certifi==2018.8.13
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.2
django-admin-sortable==2.1.4
django-appconf==1.0.2
django-extensions==2.0.7
django-imagekit==4.0.2
django-object-actions==1.0.0
django-storages==1.6.6
docutils==0.14
factory-boy==2.11.1
Faker==0.8.14
flake8==3.5.0
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
jmespath==0.9.3
mccabe==0.6.1
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
pilkit==2.0
Pillow==5.1.0
pipenv==2018.7.1
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.4
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
Unidecode==1.0.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
whitenoise==3.3.1
gunicorn==19.9.0
django-import-export==1.0.1
flake8-docstrings==1.3.0
flake8-debugger==3.1.0
